# hjyserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/device/controlLamp": {
            "post": {
                "description": "control lamp status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "controlLamp",
                "parameters": [
                    {
                        "description": "control lamp",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ControlLampReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/createStudyRoom": {
            "post": {
                "description": "create study room",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "createStudyRoom",
                "parameters": [
                    {
                        "description": "room information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.CreateStudyRoomReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StudyRoom"
                        }
                    }
                }
            }
        },
...
        "/v1/device/insert": {
            "post": {
                "description": "insert user device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "insertDevice",
                "parameters": [
                    {
                        "description": "user device information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.NewDeviceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.UserDevice"
                        }
                    }
                }
            }
        },
        "/v1/device/insertFallParams": {
            "post": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "insertFallParams",
                "parameters": [
                    {
                        "description": "fall check device params",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                }
            }
        },
        "/v1/device/modifyStudyRoom": {
            "post": {
                "description": "修改自习室信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "modifyStudyRoom",
                "parameters": [
                    {
                        "description": "room information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ModifyStudyRoomReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/device/openLampRealData": {
            "post": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "openLampRealData",
                "parameters": [
                    {
                        "description": "lamp real data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.LampRealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/queryAlarmRecord": {
            "get": {
                "description": "query alarm record report from fall check device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryAlarmRecord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallAlarm"
                        }
                    }
                }
            }
        },
        "/v1/device/queryBindByMac": {
            "get": {
                "description": "query device has bind by mac",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryBindByMac",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mdb.DeviceBindResp"
                        }
                    }
                }
            }
        },
        "/v1/device/queryById": {
            "get": {
                "description": "query device infomaion",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryById",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "device id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.Device"
                        }
                    }
                }
            }
        },
        "/v1/device/queryByUser": {
            "get": {
                "description": "query device infomaion by user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryByUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "flag",
                        "name": "flag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.Device"
                        }
                    }
                }
            }
        },
        "/v1/device/queryDateListInReport": {
            "get": {
                "description": "查詢报表中含有数据的日期列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryDateListInReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mdb.QueryDateListResp"
                        }
                    }
                }
            }
        },
        "/v1/device/queryFallCheckStatus": {
            "get": {
                "description": "query fall check data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryFallCheckStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallCheck"
                        }
                    }
                }
            }
        },
...
        "/v1/device/queryFallParams": {
            "get": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryFallParams",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "device id",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mdb.ControlLampReq": {
            "type": "object",
            "properties": {
                "brightness": {
                    "type": "integer"
                },
                "color_temp": {
                    "type": "integer"
                },
                "ctrl_mode": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "select": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "switch": {
                    "type": "integer"
                }
            }
        },
        "mdb.CreateStudyRoomReq": {
            "type": "object",
            "properties": {
                "create_id": {
                    "description": "required: true\nexample: 1",
                    "type": "integer"
                },
                "flag": {
                    "description": "required: true\nexample: 0,  1--auto invite create",
                    "type": "integer"
                },
                "name": {
                    "description": "required: true\nexample: study room",
                    "type": "string"
                }
            }
        },
...
        "mdb.ModifyStudyRoomReq": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "required: true\nexample: \"study room name\"",
                    "type": "string"
                },
                "room_id": {
                    "description": "required: true\nexample: 1",
                    "type": "integer"
                }
            }
        },
...
    }
}